+++ modules/Grades/GradeBreakdown.php
@@ -30,7 +30,7 @@ if ( ! isset( $_REQUEST['chart_type'] )
 
 echo '<form action="' . URLEscape( 'Modules.php?modname='.$_REQUEST['modname'].'' ) . '" method="GET">';
 
-$mp_select = '<select name="mp_id" id="mp_id" onchange="ajaxPostForm(this.form,true);">';
+$mp_select = '<select name="mp_id" id="mp_id" onchange="ajaxPostForm(this.form);">';
 
 foreach ( (array) $all_mp_ids as $mp_id )
 {